lab1 Finish

+++ 321labs/Game/IUnit.cs
@@ -6,8 +6,8 @@ namespace _321labs.Game
 {
     interface IUnit
     {
-        int X { get; }
-        int Y { get; }
+        double X { get; }
+        double Y { get; }
         string Name { get; }
         int Hp { get; set; }
         int Armor { get; }
